fix decoding issue
This was happening since it was decoding the url even if its not encoded
at all. It will return empty object sometimes, but sometimes it will
return object with some random unicode characters as key value pairs.

Add if condition to check if the url is encoded or not. If it is
encoded then only decode it.